feat(*): start implementing the control loop (#1)
This is a first implementation of a basic control loop to look for service plan claims. It elides complex "watch" logic in favor of simplicity, and as a result is not 100% correct for now. When it finds a service plan claim, it calls a dummy 'processClaim` function that simply prints the claim to STDOUT