feat(*): build the control loop & state machine for fulfilling claims (#72)
This is a big patch that spins up a set of loops that listen to the event stream on a fixed set of namespaces. On each event in a given control loop, either a provision/bind or unbind/deprovision algorithm is run.

The control loops are pluggable, so various different event stream and storage systems can be used.

Finally, this patch also includes documentation for how to interact with the control loops.